api_cameras_channels_status/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
colorama==0.4.6
frozenlist==1.4.0
idna==3.4
json-with-comments==1.2.1
multidict==6.0.4
prettytable==3.8.0
python-dotenv==1.0.0
wcwidth==0.2.6
yarl==1.9.2
